Update SkillManager.addSkillXp signature from long to double and update PlayerProfile XP storage map to double so fractional XP values (e.g. (#2471)
By Jarvis agent Vega.

+++ plugin/src/main/java/com/skyblock/plugin/profile/SkyBlockProfileRepository.java
@@ -54,7 +54,7 @@ public final class SkyBlockProfileRepository {
         cfg.set("showCollectionNotifications", profile.isShowCollectionNotifications());
         cfg.set("showPetNotifications", profile.isShowPetNotifications());
 
-        for (Map.Entry<String, Long> entry : profile.getSkillXp().entrySet()) {
+        for (Map.Entry<String, Double> entry : profile.getSkillXp().entrySet()) {
             cfg.set("skills." + entry.getKey(), entry.getValue());
         }
         for (Map.Entry<String, Long> entry : profile.getCollectionXp().entrySet()) {
@@ -115,7 +115,7 @@ public final class SkyBlockProfileRepository {
                     ConfigurationSection skills = cfg.getConfigurationSection("skills");
                     if (skills != null) {
                         for (String skill : skills.getKeys(false)) {
-                            profile.setSkillXp(skill, skills.getLong(skill));
+                            profile.setSkillXp(skill, skills.getDouble(skill));
                         }
                     }
